refactor: split monolithic order pipeline into multiple services
